# add Vectorization implementation for GPU