--- a/src/Distances.jl
+++ b/src/Distances.jl
@@ -1,6 +1,5 @@
 module Distances
 
-using ArrayInterface: device, AbstractDevice, GPU
 using LinearAlgebra
 using Statistics
 import StatsAPI: pairwise, pairwise!