One might even come up with a threaded `reduce` version for this, but that definitely belongs to future work.

```julia
@time using Distances
# PR: 0.284294 seconds (232.28 k allocations: 14.743 MiB, 30.94% compilation time)
# master: 0.051195 seconds (102.74 k allocations: 6.502 MiB, 85.42% compilation time)
using CUDA

X_cu = CUDA.randn(2^20);
Y_cu = CUDA.randn(2^20);

X = randn(2^20);
Y = randn(2^20);

@btime sqeuclidean($X, $Y)
# PR: 444.933 μs (0 allocations: 0 bytes)
# master: 418.690 μs (0 allocations: 0 bytes)
# almost a benchmark noisy

@btime sqeuclidean($X_cu, $Y_cu)
# GTX 3090
# PR: 51.827 μs (207 allocations: 5.64 KiB)
# master: take decades because of the scalar indexing
```

I did not expose this trait dispatching to the API level because I'm not sure if the user really wants this. Users might only want to control the computational device(e.g., GPU, CPU, CPU threads, CPU distributed) and may not want to control the under the hood implementation details.

---

> from @KristofferC in https://github.com/JuliaStats/Distances.jl/issues/143#issuecomment-522987543
> > Requires.jl makes loading this package 20x slower (#123 (comment)).

I still feel the runtime benefit worths it; not to mention that GPU support is basically unavailable before. Given that a lot of packages depend on `ArrayInterface`, adding this dependency won't add much of the loading time in any real project. (From [juliahub](https://juliahub.com/ui/Packages) results: `ArrayInterface` is depended by 692 packages while `Distances` is defended by 474 packages.)

```julia
julia> @time using ArrayInterface
  0.228708 seconds (224.97 k allocations: 13.884 MiB, 26.09% compilation time)

julia> @time using Distances
  0.014110 seconds (6.62 k allocations: 604.219 KiB)
```

Edit:

> From @nalimilan in https://github.com/JuliaStats/Distances.jl/issues/143#issuecomment-690931697
> > One solution to this would be to choose the best algorithm based on traits like those provided by ArrayInterface.

ArrayInterface requires Julia 1.2. I'll let you guys decide whether 1) we want to instead depend on Requires, or 2) directly bump the Julia lower bound here and drop Julia 1.0 and 1.1 support, or even 3) make a new package DistancesCUDA :(

closes #137