@KristofferC 2021-06-12
A quite significant effort has been made to make adding this package as a dependency a no brainer by making it have few dependencies and thus a fast load time. Unfortunately, looking at the dependency here, a 10x load time increase is observed:


```
julia> @time using ArrayInterface
  0.184645 seconds (143.03 k allocations: 9.379 MiB, 2.25% compilation time)
```

```
julia> @time using Distances
  0.017180 seconds (18.09 k allocations: 1.566 MiB, 30.99% compilation time)
```

Some possible suggestions:

- Is it possible to profile `ArrayInterface` and make its load time in the same ballpark as Distances?
- Is it possible to implement GPU support using standard Julia constructs like broadcasting?
- Perhaps this code should go into one of the GPU libraries to not make everyone absorb this increased load time.