@johnnychen94 2021-06-12
> Is it possible to profile ArrayInterface and make its load time in the same ballpark as Distances?

No, as far as I've understood. `ArrayInterface` heavily requires `Requires`

> Is it possible to implement GPU support using standard Julia constructs like broadcasting?

Yes, if you are referring to the vectorization version I just added here. But it's still slower in CPU due to the additional memory allocation for intermediate results.

> Perhaps this code should go into one of the GPU libraries to not make everyone absorb this increased load time.

The only place that requires `ArrayInterface` is `infer_evaluate_strategy`, we could only provide a fallback version here to unconditionally redirect to `ScalarMapReduce`, i.e., `infer_evaluate_strategy(d::PreMetric, a, b) = ScalarMapReduce()`. Then extends this method in other packages (e.g., DistancesCUDA.jl)

But I really don't think there's a need to be so conservative. For example, if we introduce LoopVectorization here and bring CPU runtime say >1.2x faster, does the overhead in loading time really matters? (Note that `LoopVectorization` depends on `ArrayInterface`)