@KristofferC 2021-06-12
Looking a bit more at ArrayInterface it feels like it is way out of scope to add it to something so stable and lean as Distances.jl. For example, it does a lot of type piracy. I think it makes sense to add this to one of the packages that are in the "ArrayInterface" ecosystem (with possible modifications to Distance.jl to allow that to happen).