@nalimilan 2021-06-12
It's too bad that ArrayInterface uses Requires.jl. I wish it would be a very lightweight package that others depend on, rather than the contrary. Or move the basic definitions to ArrayInterfaceBase so that packages can depend on that.