@johnnychen94 2021-06-12
> Or move the basic definitions to ArrayInterfaceBase so that packages can depend on that.

I prefer to have something like `DistancesEx.jl` in JuliaStats that allows heavier dependency for speed and glue purposes. Possible efforts to bring LoopVectorization support can be made there as well.

Or maybe we can just work in the same Distances repo `/DistancesEx` as a sub-dir package like [SnoopCompile](https://github.com/timholy/SnoopCompile.jl) does.