@devmotion 2021-06-12
If the main goal is GPU support, maybe one could also just create a GPU-specific DistancesCUDA package, similar to https://github.com/FluxML/NNlibCUDA.jl, with implementations of `_evaluate(::UnionMetrics, a::CUDA.CuArray, b::CUDA.CuArray)` etc.