@johnnychen94 2021-06-14
Close it as ArrayInterface is too heavy a dependency for Distances. Without using `ArrayInterface`, it would be better to work out a `DistancesCUDA.jl` package with tweaked CUDA kernels.